Wrote alignGap_lowMem. Removed original alignGap and constGap. Modified constGap_lowMem and view_test accordingly. Passes local and whole-repository "go test"
